Stop passing context from QuickSettingsTile and from Dialog, because we can't control order of initialization, and the same observable which needs a context is shared between QuickSettingsTile and Dialog, so the application base context makes more sense. Also the signatures of Frontlight are cleaner now.
As a good side-effect, throttling was also implemented. It helped prevent race conditions where multiple updates were initiated before the warm/cold settings got the chance to get updated.

getNamedWarmthBrightnessOptionsWithOnyxSliderSelected contained a time bomb in the form of modifying the array it was receiving. Time bomb exploded once I started calling this method with non-init data, that was supposed to not be mutated.

Added toString methods to aid with logging and later with unit test failure reports.

LifecycleAwareSubscription used instead of LiveData because I couldn't find an easy way to use LiveData without switching to AppCompat, which is not needed.