As per Android docs https://developer.android.com/reference/android/Manifest.permission#WRITE_SETTINGS this could be one cause why GentleGlow was failing to start on Android 11.
Onyx's hasCTM check replaced with check on warm & cold lights. This way, the check can be performed *before* the WRITE_SETTINGS permission check. It wouldn't make sense to ask users of e.g. Samsung tablets to first grant Modify Settings only then to be told that Gentle Glow is not for their device.