#13 ♻ Frontlight -> OnyxLight
There's going to be a distinction between Light, the proper way of doing things that Gentle Glow proposes as its main interface, and NativeLight, for which multiple interfaces may be introduced in the future. Likebooks, I'm looking at you!

Light is going to be introduced in the following commits to support unit testing.