#4 #13 prevent light off caused by rounding down both warm and cold when setting warmth in [25, 75] at brightness 1.
As a bonus of the new output correction, tougher constraints can be imposed on the whole range of values for brightness and warmth.

Users may find some of their profiles map to a slightly different warm, cold led output, and when that happens to their last selected profile, then Onyx Slider will be selected just once after the upgrade. I consider this issue to minor to justify the complexity of a migration step.